ref: move to partial class
