+++ src/Sentry.Log4Net/SentryAppender.cs
@@ -5,7 +5,7 @@ namespace Sentry.Log4Net;
 /// <summary>
 /// Sentry appender for log4net.
 /// </summary>
-public class SentryAppender : AppenderSkeleton
+public partial class SentryAppender : AppenderSkeleton
 {
     private readonly Func<string, IDisposable> _initAction;
     private volatile IDisposable? _sdkHandle;
@@ -93,31 +93,7 @@ public class SentryAppender : AppenderSkeleton
         var options = _hub.GetSentryOptions();
         if (options is { EnableLogs: true })
         {
-            var level = loggingEvent.ToSentryLogLevel();
-            if (level.HasValue)
-            {
-                DateTimeOffset timestamp = new(loggingEvent.TimeStampUtc);
-                string? template = null; // cannot get format-string from `log4net.Util.SystemStringFormat` via `LoggingEvent.MessageObject`
-                var parameters = ImmutableArray<KeyValuePair<string, object>>.Empty; // cannot get arguments from `log4net.Util.SystemStringFormat` via `LoggingEvent.MessageObject`
-
-                var log = SentryLog.Create(_hub, timestamp, level.Value, loggingEvent.RenderedMessage, template, parameters);
-
-                log.SetDefaultAttributes(options, Sdk);
-                log.SetOrigin("auto.log.log4net");
-
-                foreach (var property in loggingEvent.GetProperties())
-                {
-                    if (property is DictionaryEntry { Key: string key, Value: { } value })
-                    {
-                        if (key.Length != 0 && !key.StartsWith("log4net:", StringComparison.OrdinalIgnoreCase))
-                        {
-                            log.SetAttribute($"property.{key}", value);
-                        }
-                    }
-                }
-
-                _hub.Logger.CaptureLog(log);
-            }
+            CaptureStructuredLog(_hub, options, loggingEvent);
         }
 
         var exception = loggingEvent.ExceptionObject ?? loggingEvent.MessageObject as Exception;
